Fix Links, Header Change, Title Change
Updated the page title to simply be index.

Updated the name of the Header (which to most people will look like the title of the page) to follow the Directory Title format I have been using elsewhere.

Updated the links to just include the journals. The recommendation reports will be linked in the journals.